글쓰기를 하면 GenFile 테이블에 값이 들어가도록 구현
첨부파일에서 가져온 값이 아닌 저장 확인을 위한 코드 작성

+++ src/main/java/com/ll/social/app/fileupload/entity/GenFile.java
@@ -17,7 +17,7 @@ import javax.persistence.Entity;
 public class GenFile extends BaseEntity {
     private String relTypeCode;
 
-    private int relId;
+    private long relId;
 
     private String typeCode;
